+++ src/apps/inverse-firm/contracts/ethers/factories/index.ts
@@ -3,5 +3,8 @@
 /* eslint-disable */
 export { BorrowController__factory } from "./BorrowController__factory";
 export { Dbr__factory } from "./Dbr__factory";
+export { DbrDistributor__factory } from "./DbrDistributor__factory";
+export { RewardableEscrow__factory } from "./RewardableEscrow__factory";
 export { SimpleEscrow__factory } from "./SimpleEscrow__factory";
 export { SimpleMarket__factory } from "./SimpleMarket__factory";
+export { Xinv__factory } from "./Xinv__factory";