+++ src/apps/inverse-firm/ethereum/inverse-firm.loan.contract-position-fetcher.ts
@@ -10,17 +10,25 @@ import { DefaultDataProps } from '~position/display.interface';
 import { MetaType } from '~position/position.interface';
 import { ContractPositionTemplatePositionFetcher } from '~position/template/contract-position.template.position-fetcher';
 import {
+  GetDataPropsParams,
   GetDisplayPropsParams,
   GetTokenBalancesParams,
   GetTokenDefinitionsParams,
 } from '~position/template/contract-position.template.types';
 
 import { InverseFirmContractFactory, SimpleMarket } from '../contracts';
+import { isClaimable } from '~position/position.utils';
+
+const DBR = '0xad038eb671c44b853887a7e32528fab35dc5d710';
+const REWARD_TOKENS = {
+  '0xb516247596ca36bf32876199fbdcad6b3322330b': [DBR],
+}
 
 export type InverseFirmLoanContractPositionDefinition = {
   address: string;
   suppliedTokenAddress: string;
   borrowedTokenAddress: string;
+  rewardTokenAddresses: string[];
 };
 
 @PositionTemplate()
@@ -30,6 +38,11 @@ export class EthereumInverseFirmLoanContractPositionFetcher extends ContractPosi
   InverseFirmLoanContractPositionDefinition
 > {
   groupLabel = 'Lending';
+  dbrAddress = DBR;
+  dbrDistributor = '0xdcd2D918511Ba39F2872EB731BB88681AE184244'
+  invAddress = '0x41D5D79431A913C4aE7d69a668ecdfE5fF9DFB68'
+  xinvAddress = '0x1637e4e9941D55703a7A5E7807d6aDA3f7DCD61B'
+  dolaAddress = '0x865377367054516e17014ccded1e7d814edc9ce4'
 
   constructor(
     @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
@@ -46,7 +59,7 @@ export class EthereumInverseFirmLoanContractPositionFetcher extends ContractPosi
     const multicall = this.appToolkit.getMulticall(this.network);
 
     const dolaBorrowRightContract = this.contractFactory.dbr({
-      address: '0xad038eb671c44b853887a7e32528fab35dc5d710',
+      address: this.dbrAddress,
       network: this.network,
     });
 
@@ -61,7 +74,8 @@ export class EthereumInverseFirmLoanContractPositionFetcher extends ContractPosi
         return {
           address,
           suppliedTokenAddress: collateralTokenAddressRaw.toLowerCase(),
-          borrowedTokenAddress: '0x865377367054516e17014ccded1e7d814edc9ce4', // dola
+          borrowedTokenAddress: this.dolaAddress, // dola
+          rewardTokenAddresses: REWARD_TOKENS[address.toLowerCase()] || [],
         };
       }),
     );
@@ -83,6 +97,13 @@ export class EthereumInverseFirmLoanContractPositionFetcher extends ContractPosi
         address: definition.borrowedTokenAddress,
         network: this.network,
       },
+      ...definition.rewardTokenAddresses.map(rewardTokenAddress => {
+        return {
+          metaType: MetaType.CLAIMABLE,
+          address: rewardTokenAddress,
+          network: this.network,
+        }
+      }),
     ];
   }
 
@@ -93,19 +114,89 @@ export class EthereumInverseFirmLoanContractPositionFetcher extends ContractPosi
   async getTokenBalancesPerPosition({
     address,
     contract,
+    contractPosition,
     multicall,
   }: GetTokenBalancesParams<SimpleMarket>): Promise<BigNumberish[]> {
     const personalEscrow = await contract.escrows(address);
-    if (personalEscrow === ZERO_ADDRESS) return [0, 0];
+    if (personalEscrow === ZERO_ADDRESS) return contractPosition.tokens.map(() => 0);
 
-    const simpleEscrowContract = this.contractFactory.simpleEscrow({
+    const escrowContract = this.contractFactory.rewardableEscrow({
       address: personalEscrow.toLowerCase(),
       network: this.network,
     });
-    const supplied = await multicall.wrap(simpleEscrowContract).balance();
 
+    const supplied = await multicall.wrap(escrowContract).balance();
     const borrowed = await contract.debts(address);
 
-    return [supplied, borrowed];
+    const rewardTokens = contractPosition.tokens.filter(isClaimable);
+    const rewardBalances = await Promise.all(
+      rewardTokens.map(token => {
+        if (DBR === token.address.toLowerCase()) {
+          return multicall.wrap(escrowContract).claimable()
+        }
+        return Promise.resolve(0);
+      })
+    );
+
+    return [supplied, borrowed, ...rewardBalances];
+  }
+
+  async getDataProps(
+    params: GetDataPropsParams<SimpleMarket>,
+  ): Promise<DefaultDataProps> {
+    const { contractPosition } = params;
+    const rewardToken = contractPosition.tokens.find(isClaimable)!;
+    if (!rewardToken) return {};
+
+    if (DBR === rewardToken.address.toLowerCase()) {
+      return {
+        supplyApy: await this.getDBRApr(params),
+      }
+    }
+
+    return {};
+  }
+
+  async getDBRApr({
+    multicall
+  }) {
+    const dbrDistributorContract = this.contractFactory.dbrDistributor({
+      address: this.dbrDistributor,
+      network: this.network,
+    });
+    const xinvContract = this.contractFactory.xinv({
+      address: this.xinvAddress,
+      network: this.network,
+    });
+
+    const [
+      rewardRate,
+      dbrDistributorSupply,
+      xinvExRate,
+    ] = await Promise.all([
+      multicall.wrap(dbrDistributorContract).rewardRate(),
+      multicall.wrap(dbrDistributorContract).totalSupply(),
+      multicall.wrap(xinvContract).exchangeRateStored(),
+    ]);
+
+    const [dbrtoken, invToken] = await Promise.all([
+      this.appToolkit.getBaseTokenPrice({
+        network: this.network,
+        address: this.dbrAddress,
+      }),
+      this.appToolkit.getBaseTokenPrice({
+        network: this.network,
+        address: this.invAddress,
+      }),
+    ]);
+
+    if (!dbrtoken || !invToken) return [0];
+
+    const invStakedViaDistributor = xinvExRate.mul(dbrDistributorSupply);
+    const dbrYearlyRewardRate = rewardRate.mul(31536000);
+    const dbrInvExRate = dbrtoken.price / invToken.price;
+    const dbrApr = Number(dbrYearlyRewardRate) * Number(dbrInvExRate) / Number(invStakedViaDistributor) / 1e18 * 100;
+
+    return dbrApr;
   }
 }